test-web: Move test-collection code into its own files
No behaviour changes.
