+++ Tests/LibWeb/test-web/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(SOURCES
     Application.cpp
     CaptureFile.cpp
+    Collection.cpp
     Debug.cpp
     Display.cpp
     Fixture.cpp
